Move enum & interfaces to configs/
