+++ ui/InputReceiver.java
@@ -1,9 +1,10 @@
 package ui;
 
-import java.util.*;
+import java.util.Scanner;
 import algorithm.*;
 import control.*;
 import models.*;
+import configs.UnitType;
 
 // 入力受け付けクラス
 public class InputReceiver {
